// house-rental/src/app/new-house-form/models/manager.model.ts
export interface User {
    id: number;
    name: string;
    email: string;
    phone: string;
    role: string;
    position: string;
    status: string;
    houseNames: string[];
    birthDate: string;
    gender: string;
    startDate: string;
    idNumber: string;
    description: string;
  }
  